+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/ui/screens/main/components/ConnectionButton.kt
@@ -34,6 +34,7 @@ fun ConnectionButton(
 	connectionState: ConnectionState,
 	accountState: AccountControllerState,
 	isMnemonicStored: Boolean,
+	isAccountInitializing: Boolean,
 	onConnect: () -> Unit,
 	onDisconnect: () -> Unit,
 	onStopKillSwitch: () -> Unit,